Refactoring, fixed the issues the last build caused

+++ controllers/user_controller.py
@@ -1,5 +1,6 @@
 from models.user import User 
 from mysql.connector import Error
+from sqlite3 import Error
 import time
 
 class UserController:
@@ -24,7 +25,7 @@ class UserController:
         user = User(name, password)
         return user
         
-    def create_new_user(self):
+    def create_user(self):
         new_user = self.get_user_info()
         sql = f"INSERT INTO users (userId, name, password) VALUES (\'{new_user.id}\', \'{new_user.user_name}\', \'{new_user.password}\')"
         try:
@@ -32,10 +33,11 @@ class UserController:
             self.connection.commit()
             print(self.cur.rowcount, "record inserted")
             self.db.user_cache += [(new_user.id, new_user.user_name)]
-            return 1
+            return 'user_created'
         except Error as e:
-            print(e)
-            return -4
+            return 'user_found'
+        except Exception as e:
+            return 'user_found'
         
     def display_users(self):
         try:
@@ -47,50 +49,52 @@ class UserController:
                 time.sleep(0.5)
         except Error as e:
             print(e)
-        return 4
+        [print(i) for i in self.db.user_cache]
             
     def update_user_name(self):
         try:
             new_name = input('Enter new username: ')
             if len(new_name) == 0:
-                return -2
+                return 'invalid_name'
             else:
                 self.db.update_user_cache(self._cur_user, new_name, action='PUTS')
                 self._cur_user.user_name = new_name
                 self.cur.execute(f'UPDATE users SET name = \'{new_name}\' WHERE userId = \'{self._cur_user.id}\'')
                 self.connection.commit()
                 self._cur_user.user_name = new_name
-                return 9
+                return 'user_updated'
         except Error as e:
             print(e)
             
     def delete_user(self):
-        try:
-            self.cur.execute(f'DELETE FROM users WHERE userId = \'{self._cur_user.id}\'')
-            self.connection.commit()
-            self.db.update_user_cache(self._cur_user, action='DELETE')
-            self._cur_user = None
-            return -7
-        except Error as e:
-            print(e)
+        choice = input('Are you sure? y/n ').casefold()
+        if choice == 'y':
+            try:
+                self.cur.execute(f'DELETE FROM users WHERE userId = \'{self._cur_user.id}\'')
+                self.connection.commit()
+                self.db.update_user_cache(self._cur_user, action='DELETE')
+                self._cur_user = None
+                return 'user_deleted'
+            except Error as e:
+                print(e)
     
-    def sign_in(self):
+    def sign_in(self, deck):
         check_user = self.get_user_info()
         try:
             self.cur.execute(f'SELECT userId, name, password FROM users WHERE name = \'{check_user.user_name}\'')
             row = self.cur.fetchone()
-        except Error as e:
+        except Exception as e:
             print(e)
-            return -2
-        
+            return 'invalid_choice'
         if row is None:
-            return -6
+            return 'user_not_found'
         else:
             if check_user.password != row[2]:
-                return -3
+                return 'invalid_password'
             else:
                 self._cur_user = User(id=row[0], user_name=row[1], password=row[2])
-                return 2
+                deck.cur_user = self._cur_user
+                return 'use_user'
 
     def sign_out(self, deck):
         self._cur_user = None
